backend/requirements.txt
flask
flask-pymongo
jwt
python-dotenv
flask-cors
requests